backend/requirements.txt
api==0.0.7
blis==0.4.1
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
cymem==2.0.3
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
gunicorn==20.0.4
idna==2.10
flask==1.1.2
flask_cors==3.0.9
mdutils==1.3.0
murmurhash==1.0.2
nose==1.3.7
numpy==1.19.2
plac==1.1.3
preshed==3.0.2
python-dateutil==2.8.1
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
spacy==2.3.2
srsly==1.0.2
thinc==7.4.1
tqdm==4.49.0
urllib3==1.25.10
wasabi==0.8.0
